// mfes/workspace/src/utils/interfaces.ts
export interface CohortMemberList {
 limit?: number;
 page?: number;
 filters: {
   cohortId: string;
   role?: string;
   status?: string[];
   name?: string | undefined;
   firstName?: string;
 };
 includeArchived?: boolean;
}


export interface UserList {
 limit: number;
 page: number;
 filters: {
   states: string;
   districts: string;
   blocks: string;
   role?: string;
 };
 fields: string[];
}
// ...
export interface UpdateCohortMemberStatusParams {
 memberStatus?: string;
 statusReason?: string;
 membershipId: string | number;
 dynamicBody?: Record<string, any>;
}